# Optimize type resolving