--- a/src/NJsonSchema.NewtonsoftJson.Tests/NJsonSchema.NewtonsoftJson.Tests.csproj
+++ b/src/NJsonSchema.NewtonsoftJson.Tests/NJsonSchema.NewtonsoftJson.Tests.csproj
@@ -26,7 +26,6 @@
 
     <ItemGroup>
         <ProjectReference Include="..\NJsonSchema.NewtonsoftJson\NJsonSchema.NewtonsoftJson.csproj" />
-        <ProjectReference Include="..\NJsonSchema\NJsonSchema.csproj" />
     </ItemGroup>
 
 </Project>