--- a/src/NJsonSchema.NewtonsoftJson/NJsonSchema.NewtonsoftJson.csproj
+++ b/src/NJsonSchema.NewtonsoftJson/NJsonSchema.NewtonsoftJson.csproj
@@ -9,21 +9,6 @@
         <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
     </ItemGroup>
 
-    <ItemGroup Condition="'$(TargetFramework)' == 'net462'">
-        <PackageReference Include="Microsoft.Bcl.Async">
-            <Version>1.0.168</Version>
-        </PackageReference>
-    </ItemGroup>
-
-    <ItemGroup Condition="'$(TargetFramework)' == 'net462'">
-        <Reference Include="Microsoft.CSharp" />
-        <Reference Include="System" />
-        <Reference Include="System.Net" />
-        <Reference Include="System.Runtime.Serialization" />
-        <Reference Include="System.Xml" />
-        <Reference Include="System.Xml.Linq" />
-    </ItemGroup>
-
     <ItemGroup>
         <ProjectReference Include="..\NJsonSchema\NJsonSchema.csproj" />
     </ItemGroup>