--- a/src/NJsonSchema.Yaml.Tests/NJsonSchema.Yaml.Tests.csproj
+++ b/src/NJsonSchema.Yaml.Tests/NJsonSchema.Yaml.Tests.csproj
@@ -16,7 +16,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\NJsonSchema\NJsonSchema.csproj" />
     <ProjectReference Include="..\NJsonSchema.Yaml\NJsonSchema.Yaml.csproj" />
   </ItemGroup>