* reuse same `HashSet` for `reservedTypeNames` instead re-creating on every cache miss
* cleanup csprojs
* let NUKE regenerate GH actions definitions
* remove build configurations for `_build` project in solution file (prevents local command line usage)
* use string interpolation for better optimizations (NET 6+ has optimizations)

The test case has only 3 types and shows the difference quite nicely, real world usage can have hundreds of items, if not thousands.

## NJsonSchema.Benchmark.CSharpTypeResolverBenchmark

| **Diff**|Method|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old |RegisterSchemaDefinitions|1.119 μs|0.0162 μs|1.66 KB|
| **New** |	| **645.8 ns (-44%)** | **5.40 ns** | **888 B (-48%)** |